CarFuelGreedy.java
It's a famous greedy algorithm problem. The input consists of the number of gas stations on the way (n), Max distance that can be covered per fill( maxDis), and the array of distance from the beginning point (a). The algorithm finds the minimum fills required by car to reach the destination.